backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
openpyxl
python-dotenv
openai>=1.0.0
supabase
